+++ GeoGL/gengine/image/image2d.h
@@ -6,13 +6,13 @@ namespace gengine {
 		class Image2D
 		{
 		public:
-			size_t _width;
-			size_t _height;
+			unsigned int _width;
+			unsigned int _height;
 			unsigned int* _raw_buffer;
 			Image2D(void);
 			~Image2D(void);
 		protected:
-			void InitBuffer(size_t w, size_t h);
+			void InitBuffer(unsigned int w, unsigned int h);
 		};
 
 	} //namespace image
